# Refactor CpuMathUtils