--- a/src/Microsoft.ML.CpuMath/AvxIntrinsics.cs
+++ b/src/Microsoft.ML.CpuMath/AvxIntrinsics.cs
@@ -606,12 +606,12 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe void ScaleSrcU(float scale, Span<float> src, Span<float> dst)
+        public static unsafe void ScaleSrcU(float scale, ReadOnlySpan<float> src, Span<float> dst, int count)
         {
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
-                float* pDstEnd = pdst + dst.Length;
+                float* pDstEnd = pdst + count;
                 float* pSrcCurrent = psrc;
                 float* pDstCurrent = pdst;
 
@@ -697,14 +697,14 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe void AddScaleU(float scale, Span<float> src, Span<float> dst)
+        public static unsafe void AddScaleU(float scale, ReadOnlySpan<float> src, Span<float> dst, int count)
         {
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
                 float* pSrcCurrent = psrc;
                 float* pDstCurrent = pdst;
-                float* pEnd = pdst + dst.Length;
+                float* pEnd = pdst + count;
 
                 Vector256<float> scaleVector256 = Avx.SetAllVector256(scale);
 
@@ -751,13 +751,13 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe void AddScaleCopyU(float scale, Span<float> src, Span<float> dst, Span<float> result)
+        public static unsafe void AddScaleCopyU(float scale, ReadOnlySpan<float> src, ReadOnlySpan<float> dst, Span<float> result, int count)
         {
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             fixed (float* pres = result)
             {
-                float* pResEnd = pres + result.Length;
+                float* pResEnd = pres + count;
                 float* pSrcCurrent = psrc;
                 float* pDstCurrent = pdst;
                 float* pResCurrent = pres;
@@ -807,7 +807,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe void AddScaleSU(float scale, Span<float> src, Span<int> idx, Span<float> dst)
+        public static unsafe void AddScaleSU(float scale, ReadOnlySpan<float> src, ReadOnlySpan<int> idx, Span<float> dst, int count)
         {
             fixed (float* psrc = src)
             fixed (int* pidx = idx)
@@ -816,7 +816,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                 float* pSrcCurrent = psrc;
                 int* pIdxCurrent = pidx;
                 float* pDstCurrent = pdst;
-                int* pEnd = pidx + idx.Length;
+                int* pEnd = pidx + count;
 
                 Vector256<float> scaleVector256 = Avx.SetAllVector256(scale);
 
@@ -858,14 +858,14 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe void AddU(Span<float> src, Span<float> dst)
+        public static unsafe void AddU(ReadOnlySpan<float> src, Span<float> dst, int count)
         {
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
                 float* pSrcCurrent = psrc;
                 float* pDstCurrent = pdst;
-                float* pEnd = psrc + src.Length;
+                float* pEnd = psrc + count;
 
                 while (pSrcCurrent + 8 <= pEnd)
                 {
@@ -905,7 +905,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe void AddSU(Span<float> src, Span<int> idx, Span<float> dst)
+        public static unsafe void AddSU(ReadOnlySpan<float> src, ReadOnlySpan<int> idx, Span<float> dst, int count)
         {
             fixed (float* psrc = src)
             fixed (int* pidx = idx)
@@ -914,7 +914,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                 float* pSrcCurrent = psrc;
                 int* pIdxCurrent = pidx;
                 float* pDstCurrent = pdst;
-                int* pEnd = pidx + idx.Length;
+                int* pEnd = pidx + count;
 
                 while (pIdxCurrent + 8 <= pEnd)
                 {
@@ -950,7 +950,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe void MulElementWiseU(Span<float> src1, Span<float> src2, Span<float> dst)
+        public static unsafe void MulElementWiseU(ReadOnlySpan<float> src1, ReadOnlySpan<float> src2, Span<float> dst, int count)
         {
             fixed (float* psrc1 = src1)
             fixed (float* psrc2 = src2)
@@ -959,7 +959,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                 float* pSrc1Current = psrc1;
                 float* pSrc2Current = psrc2;
                 float* pDstCurrent = pdst;
-                float* pEnd = pdst + dst.Length;
+                float* pEnd = pdst + count;
 
                 while (pDstCurrent + 8 <= pEnd)
                 {
@@ -999,7 +999,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float SumU(Span<float> src)
+        public static unsafe float SumU(ReadOnlySpan<float> src)
         {
             fixed (float* psrc = src)
             {
@@ -1037,7 +1037,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float SumSqU(Span<float> src)
+        public static unsafe float SumSqU(ReadOnlySpan<float> src)
         {
             fixed (float* psrc = src)
             {
@@ -1081,7 +1081,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float SumSqDiffU(float mean, Span<float> src)
+        public static unsafe float SumSqDiffU(float mean, ReadOnlySpan<float> src)
         {
             fixed (float* psrc = src)
             {
@@ -1130,7 +1130,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float SumAbsU(Span<float> src)
+        public static unsafe float SumAbsU(ReadOnlySpan<float> src)
         {
             fixed (float* psrc = src)
             {
@@ -1174,7 +1174,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float SumAbsDiffU(float mean, Span<float> src)
+        public static unsafe float SumAbsDiffU(float mean, ReadOnlySpan<float> src)
         {
             fixed (float* psrc = src)
             {
@@ -1223,7 +1223,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float MaxAbsU(Span<float> src)
+        public static unsafe float MaxAbsU(ReadOnlySpan<float> src)
         {
             fixed (float* psrc = src)
             {
@@ -1267,7 +1267,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float MaxAbsDiffU(float mean, Span<float> src)
+        public static unsafe float MaxAbsDiffU(float mean, ReadOnlySpan<float> src)
         {
             fixed (float* psrc = src)
             {
@@ -1316,14 +1316,14 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float DotU(Span<float> src, Span<float> dst)
+        public static unsafe float DotU(ReadOnlySpan<float> src, ReadOnlySpan<float> dst, int count)
         {
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
                 float* pSrcCurrent = psrc;
                 float* pDstCurrent = pdst;
-                float* pSrcEnd = psrc + src.Length;
+                float* pSrcEnd = psrc + count;
 
                 Vector256<float> result256 = Avx.SetZeroVector256<float>();
 
@@ -1371,7 +1371,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float DotSU(Span<float> src, Span<float> dst, Span<int> idx)
+        public static unsafe float DotSU(ReadOnlySpan<float> src, ReadOnlySpan<float> dst, ReadOnlySpan<int> idx, int count)
         {
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
@@ -1380,7 +1380,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                 float* pSrcCurrent = psrc;
                 float* pDstCurrent = pdst;
                 int* pIdxCurrent = pidx;
-                int* pIdxEnd = pidx + idx.Length;
+                int* pIdxEnd = pidx + count;
 
                 Vector256<float> result256 = Avx.SetZeroVector256<float>();
 
@@ -1428,14 +1428,14 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe float Dist2(Span<float> src, Span<float> dst)
+        public static unsafe float Dist2(ReadOnlySpan<float> src, ReadOnlySpan<float> dst, int count)
         {
             fixed (float* psrc = src)
             fixed (float* pdst = dst)
             {
                 float* pSrcCurrent = psrc;
                 float* pDstCurrent = pdst;
-                float* pSrcEnd = psrc + src.Length;
+                float* pSrcEnd = psrc + count;
 
                 Vector256<float> sqDistanceVector256 = Avx.SetZeroVector256<float>();
 
@@ -1482,13 +1482,13 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe void SdcaL1UpdateU(float primalUpdate, Span<float> src, float threshold, Span<float> v, Span<float> w)
+        public static unsafe void SdcaL1UpdateU(float primalUpdate, int count, ReadOnlySpan<float> src, float threshold, Span<float> v, Span<float> w)
         {
             fixed (float* psrc = src)
             fixed (float* pdst1 = v)
             fixed (float* pdst2 = w)
             {
-                float* pSrcEnd = psrc + src.Length;
+                float* pSrcEnd = psrc + count;
                 float* pSrcCurrent = psrc;
                 float* pDst1Current = pdst1;
                 float* pDst2Current = pdst2;
@@ -1544,14 +1544,14 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             }
         }
 
-        public static unsafe void SdcaL1UpdateSU(float primalUpdate, Span<float> src, Span<int> indices, float threshold, Span<float> v, Span<float> w)
+        public static unsafe void SdcaL1UpdateSU(float primalUpdate, int count, ReadOnlySpan<float> src, ReadOnlySpan<int> indices, float threshold, Span<float> v, Span<float> w)
         {
             fixed (float* psrc = src)
             fixed (int* pidx = indices)
             fixed (float* pdst1 = v)
             fixed (float* pdst2 = w)
             {
-                int* pIdxEnd = pidx + indices.Length;
+                int* pIdxEnd = pidx + count;
                 float* pSrcCurrent = psrc;
                 int* pIdxCurrent = pidx;