… src/Microsoft.ML.CpuMath/CpuMathUtils.netcoreapp.cs
… src/Microsoft.ML.CpuMath/CpuMathUtils.netstandard.cs
… src/Microsoft.ML.CpuMath/Sse.cs
… src/Microsoft.ML.CpuMath/SseIntrinsics.cs
… src/Microsoft.ML.Data/Depricated/Vector/VectorUtils.cs
… src/Microsoft.ML.StandardLearners/Standard/SdcaMultiClass.cs
… src/Microsoft.ML.Transforms/GcnTransform.cs
… test/Microsoft.ML.CpuMath.PerformanceTests/AvxPerformanceTests.cs
… test/Microsoft.ML.CpuMath.PerformanceTests/SsePerformanceTests.cs
… test/Microsoft.ML.CpuMath.UnitTests.netcoreapp/UnitTests.cs