@eerhardt 2018-10-15
Here are some results from my machine:

Before:
```
              Method |         Mean |      Error |     StdDev |        Extra Metric |
-------------------- |-------------:|-----------:|-----------:|--------------------:|
    TrainKMeansAndLR |     709.1 ms |   44.86 ms |   49.86 ms |                   - |
           TrainIris |     262.9 ms |   12.43 ms |   14.31 ms |                   - |
      TrainSentiment | 1,763.720 ms | 22.2524 ms | 20.8149 ms |                   - |
         PredictIris |     1.648 ms |  0.0104 ms |  0.0087 ms | AccuracyMacro: 0.98 |
 PredictIrisBatchOf1 |     1.623 ms |  0.0361 ms |  0.0416 ms | AccuracyMacro: 0.98 |
 PredictIrisBatchOf2 |     1.631 ms |  0.0391 ms |  0.0450 ms | AccuracyMacro: 0.98 |
 PredictIrisBatchOf5 |     1.592 ms |  0.0326 ms |  0.0375 ms | AccuracyMacro: 0.98 |
```

After:
```
              Method |         Mean |      Error |     StdDev |        Extra Metric |
-------------------- |-------------:|-----------:|-----------:|--------------------:|
    TrainKMeansAndLR |     712.8 ms |   48.01 ms |   53.37 ms |                   - |
           TrainIris |     266.7 ms |   9.616 ms |   10.69 ms |                   - |
      TrainSentiment | 1,765.497 ms | 37.6166 ms | 41.8108 ms |                   - |
         PredictIris |     1.656 ms |  0.0191 ms |  0.0170 ms | AccuracyMacro: 0.98 |
 PredictIrisBatchOf1 |     1.546 ms |  0.0276 ms |  0.0258 ms | AccuracyMacro: 0.98 |
 PredictIrisBatchOf2 |     1.645 ms |  0.0366 ms |  0.0421 ms | AccuracyMacro: 0.98 |
 PredictIrisBatchOf5 |     1.574 ms |  0.0407 ms |  0.0468 ms | AccuracyMacro: 0.98 |
```

Without the calls to `MemoryMarshal.GetReference` when pinning the Span, I am seeing:


```
              Method |         Mean |      Error |     StdDev |        Extra Metric |
-------------------- |-------------:|-----------:|-----------:|--------------------:|
    TrainKMeansAndLR |     810.2 ms |   71.12 ms |   79.05 ms |                   - |
           TrainIris |   260.214 ms |  9.7503 ms | 11.2285 ms |                   - |
      TrainSentiment | 1,752.501 ms | 24.0648 ms | 22.5103 ms |                   - |
         PredictIris |     1.650 ms |  0.0151 ms |  0.0141 ms | AccuracyMacro: 0.98 |
 PredictIrisBatchOf1 |     1.599 ms |  0.0362 ms |  0.0339 ms | AccuracyMacro: 0.98 |
 PredictIrisBatchOf2 |     1.646 ms |  0.0179 ms |  0.0167 ms | AccuracyMacro: 0.98 |
 PredictIrisBatchOf5 |     1.635 ms |  0.0192 ms |  0.0179 ms | AccuracyMacro: 0.98 |
```

As you can see, the `TrainKMeansAndLR` benchmark was really affected by the change to `MemoryMarshal.GetReference`.