@Anipik 2018-10-15
@eerhardt wont it be nice to also  check the performance directly using cpumath.performaceTests
like https://github.com/dotnet/machinelearning/blob/0b84350571fd0c272faa03541964c0d0617e0d76/test/Microsoft.ML.CpuMath.PerformanceTests/AvxPerformanceTests.cs#L15