@eerhardt 2018-10-15
> wont it be nice to also check the performance directly using cpumath.performaceTests
like

I only ran the `Avx.Sum*` perf tests because these take so long.

Before:

```
      Method |     Mean |    Error |   StdDev |
------------ |---------:|---------:|---------:|
        SumU | 205.2 us | 4.053 us | 4.668 us |
      SumSqU | 220.6 us | 4.096 us | 4.023 us |
  SumSqDiffU | 230.4 us | 4.471 us | 5.814 us |
     SumAbsU | 233.6 us | 4.612 us | 8.317 us |
 SumAbsDiffU | 240.9 us | 4.775 us | 8.237 us |
```

After:
```
      Method |     Mean |    Error |   StdDev |
------------ |---------:|---------:|---------:|
        SumU | 197.5 us | 3.030 us | 2.686 us |
      SumSqU | 213.3 us | 4.000 us | 3.546 us |
  SumSqDiffU | 218.7 us | 2.469 us | 2.309 us |
     SumAbsU | 240.8 us | 2.979 us | 2.641 us |
 SumAbsDiffU | 233.8 us | 8.379 us | 8.966 us |
```